modify reset_lgw.sh to slow down

+++ reset_lgw.sh
@@ -8,10 +8,10 @@
 #       ./reset_lgw.sh stop
 #       ./reset_lgw.sh start
 
-# The reset pin of SX1301 is wired with RPi GPIO7
+# The reset pin connected to Nitrogen8m is GPIO103
 # If used on another platform, the GPIO number can be given as parameter.
 if [ -z "$2" ]; then 
-    IOT_SK_SX1301_RESET_PIN=7
+    IOT_SK_SX1301_RESET_PIN=103
 else
     IOT_SK_SX1301_RESET_PIN=$2
 fi
@@ -24,16 +24,27 @@ WAIT_GPIO() {
 
 iot_sk_init() {
     # setup GPIO 7
-    echo "$IOT_SK_SX1301_RESET_PIN" > /sys/class/gpio/export; WAIT_GPIO
+    file=/sys/class/gpio/gpio$IOT_SK_SX1301_RESET_PIN
+    if [ ! \( -e "${file}" \) ]
+    then
+       echo "GPIO needs to be exported."
+       echo "$IOT_SK_SX1301_RESET_PIN" > /sys/class/gpio/export; WAIT_GPIO
+       sleep 2
+    else
+       echo "GPIO already exported."
+    fi
 
     # set GPIO 7 as output
+    echo "set direction to output"
     echo "out" > /sys/class/gpio/gpio$IOT_SK_SX1301_RESET_PIN/direction; WAIT_GPIO
 
     # write output for SX1301 reset
+    echo "toggle high then low."
     echo "1" > /sys/class/gpio/gpio$IOT_SK_SX1301_RESET_PIN/value; WAIT_GPIO
     echo "0" > /sys/class/gpio/gpio$IOT_SK_SX1301_RESET_PIN/value; WAIT_GPIO
 
     # set GPIO 7 as input
+    echo "set direction to input"
     echo "in" > /sys/class/gpio/gpio$IOT_SK_SX1301_RESET_PIN/direction; WAIT_GPIO
 }
 
@@ -47,7 +58,6 @@ iot_sk_term() {
 
 case "$1" in
     start)
-    iot_sk_term
     iot_sk_init
     ;;
     stop)
